third_party/WebKit/LayoutTests/inspector/sources/debugger/js-with-inline-stylesheets.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>

<script>

function testFunction()
{
    var x = Math.sqrt(10);
    return x;
}

var test = function() {
  var currentSourceFrame;
  SourcesTestRunner.setQuiet(true);
  var pageURL = 'js-with-inline-stylesheets.html';
  SourcesTestRunner.runDebuggerTestSuite([function testSetBreakpoint(next) {
    SourcesTestRunner.showScriptSource(pageURL, didShowScriptSource);

    function didShowScriptSource(sourceFrame) {
      currentSourceFrame = sourceFrame;
      TestRunner.addResult('Script source was shown.');
      SourcesTestRunner.waitUntilPaused(didPause);
      SourcesTestRunner.createNewBreakpoint(currentSourceFrame, 9, '', true)
          .then(() => SourcesTestRunner.waitBreakpointSidebarPane())
          .then(() => SourcesTestRunner.runTestFunction());
    }

    function didPause(callFrames) {
      TestRunner.addResult('Script execution paused.');
      SourcesTestRunner.captureStackTrace(callFrames);
      SourcesTestRunner.dumpBreakpointSidebarPane();
      SourcesTestRunner.waitBreakpointSidebarPane().then(breakpointRemoved);
      SourcesTestRunner.removeBreakpoint(currentSourceFrame, 9);
    }

    function breakpointRemoved() {
      SourcesTestRunner.resumeExecution(didResume);
    }

    function didResume() {
      SourcesTestRunner.dumpBreakpointSidebarPane();
      TestRunner.addResult('Script execution resumed.');
      next();
    }
  }]);
};

</script>
<style>
body {
    color: green;
}
</style>

</head>

<body onload="runTest()">
<p>
Tests that JS sourcemapping for inline scripts followed by inline stylesheets does not break.
</p>

</body>
</html>
